+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- The running version is shown in the user menu, under Logout. If the frontend
+  and backend are somehow on different versions -- which is what a half-finished
+  deploy looks like -- both are shown rather than one being picked.
+
 ## [2.1.0-beta.6] - 2026-09-05
 
 ### Added